Size "mediumint" wasn't enough: https://digitalcourage.video/static/webseed/5a039eb2-9fbc-441d-a772-59ae3ee65c15-1080.mp4 has 6758472669 Bytes
So let's go REALLY big.

+++ database.sql
@@ -1221,7 +1221,7 @@ CREATE TABLE IF NOT EXISTS `post-media` (
 	`mimetype` varchar(60) COMMENT '',
 	`height` smallint unsigned COMMENT 'Height of the media',
 	`width` smallint unsigned COMMENT 'Width of the media',
-	`size` mediumint unsigned COMMENT 'Media size',
+	`size` bigint unsigned COMMENT 'Media size',
 	`preview` varbinary(255) COMMENT 'Preview URL',
 	`preview-height` smallint unsigned COMMENT 'Height of the preview picture',
 	`preview-width` smallint unsigned COMMENT 'Width of the preview picture',
